"SeekableBzip2File" that allows semi-random access to bzip2 compressed files by using a block position table as produced by `bzip-table`.
Consecutive reads do not require any seeking or reseting of the decompression
buffers, et cetera. That only needs to be done if a seek is performed. Seeking
involves finding the compressed block a position belongs to (blocks will be
about 200-900k on disk, but could contain much more uncompressed data). The
block will then be decompressed up to the requested position. There are some
optimizations that will help here (optimizing forward seeks in the same block,
these do not require reading from disk since the data is already in the
decompression buffer, but they currently do).

For caching, given that an uncompressed block can contain an arbitrary amount
of data, it probably makes the most sense to implement an arbitrary chunk
cache for file-like objects on top of this. Another options might be to cache
the decompression buffers (after huffman decoding but before BWT+RLE). Could
be a mess though. Since a block can be huge when decompressed (46gigsish if
completely repetitive for example) caching whole blocks is not really
reasonable.

Tests are currently commented out since it needs the files generate by
"make test" in the seek-bunzip tree. If you run that you can enable the tests.