Fix memory issues when dealing with alignments.
Make component.alignment a weakref and change name to component._alignment. Add __deepcopy__ methods to both component and alignment (used in fusing alignments), as weakref is not compatible with python's default behavior.

Add an iterator version of index.get() for returning blocks (index offsets are still processed as an array, to allow sorting).